+++ internal/tools/tools.go
@@ -21,7 +21,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package signaling
+package tools
 
 // Import applications that would otherwise not be detected by "go mod vendor".
 import (
