+++ internal/tools/vendor_helper_test.go
@@ -19,7 +19,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package signaling
+package tools
 
 // Import modules that would otherwise not be detected by "go mod vendor".
 import (
